+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 11/4/2023
+- Minor: Reworked new code checker and codes now check for hoyolab if there's any new codes too
+
 ## 10/29/2023
 - Major: Writed newly obtained redeem codes to `codes.js` file to avoid duplicate codes